fix(tx-pool): RBF robustness, bounded retries, queue budgets and structural cleanup
Security and robustness:
- RBF: pre-validate ancestor capacity before removing conflicts, so a
  replacement certain to fail cannot churn the pool with repeated
  evict-then-restore cycles at no cost
- RBF: compute the conflict closure in a single multi-source traversal
  shared by rule checks, commit pre-validation and removal, replacing
  per-conflict descendant walks; the candidate cap is now enforced
  inside the traversal so an oversized union costs
  O(MAX_REPLACEMENT_CANDIDATES) regardless of pool population
- ordered resolve: bound in-flight orphan retries
  (MAX_LOCAL_ORPHAN_IN_FLIGHT_ATTEMPTS, 2400 in production) and move
  them from spawn-per-retry tasks to a delayed heap inside the queue,
  keeping delayed jobs visible to remove_tx and RPC at all times
- verify manager: add a monitor with panic-respawn, and exponential
  backoff (100ms doubling, 30s cap, healthy-run reset) shared by the
  verify manager, ordered resolver and pre-check worker respawn loops
- commit_entry_to_pool: turn the silent short-id duplicate outcome
  into an error so removed conflicts go through the normal recovery
  path instead of being silently dropped

Performance and configuration:
- verify queue budget is now configurable via
  max_verify_queue_tx_size (default 256MB, clamped up to
  max_tx_pool_size with a warning so persisted-pool reload cannot hit
  Reject::Full); pre-check and ordered resolve queue budgets tuned to
  64MB each
- all_missing_parents_in_flight: hoist queue read locks out of the
  per-parent loop
- get_tx_detail: compute rank/counts from the sorted index and pool
  stats instead of allocating full id vectors per query

Structure (no behavior change):
- split pool/rbf.rs, process/classify.rs, process/remove.rs and
  block_assembler/{cell_liveness,dao}.rs out of their host files
- consolidate query helpers into service/dispatch.rs
- unify production and benchmark service construction behind
  assemble_service (the previous "single source of truth" doc was
  inaccurate; it is now real)
- consolidate the six duplicated test service constructions into one
  harness builder

Tests: direct unit tests for the ordered-queue delay model, conflict
closure limit abort, RBF commit pre-validation, retry bounds and
respawn backoff. Full suite passes (cargo nextest run -p ckb-tx-pool
--lib, 107 tests) and make clippy is clean.

+++ tx-pool/src/resolved_tx.rs
@@ -23,8 +23,10 @@ pub struct ResolveJob {
     pub source: TxSource,
     /// Number of times this local transaction has been retried because its
     /// inputs were not yet available. Used to bound retries for orphans that
-    /// are not satisfiable.
-    pub attempts: u8,
+    /// are not satisfiable (`MAX_LOCAL_ORPHAN_ATTEMPTS`) and for orphans whose
+    /// parents stay in flight indefinitely
+    /// (`MAX_LOCAL_ORPHAN_IN_FLIGHT_ATTEMPTS`).
+    pub attempts: u16,
 }
 
 impl ResolveJob {
